Added driveFeet and drive to DriveTrain.java
driveFeet(double) and drive(double, double) are functions for the autonomous sequence which take in feet, and feet and inches, respectively. Both derive directly from driveInches(double).
It's not a necessary addition, and they might never be used, but it's just to give ourselves a little breathing room. Nothing wrong with that.